back-end/requirements.txt
fastapi>=0.68.0
pydantic>=1.8.0
uvicorn>=0.15.0
pymongo>=4.2.0
pydantic[email]
motor>=2.5.1
python-dateutil>=2.8.2

